# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for Campwiz",
        "title": "Campwiz API",
        "contact": {
            "name": "Nokib Sarkar",
            "url": "https://github.com/nokibsarkar",
            "email": "nokibsarkar@gmail.com"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "1"
    },
    "host": "localhost:8081",
    "basePath": "/api/v2",
    "paths": {
...
        "/submission/": {
            "get": {
                "description": "get all submissions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Submission"
                ],
                "summary": "List all submissions",
                "parameters": [
                    {
                        "type": "string",
                        "name": "campaignId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "next",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "participantId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "prev",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "roundId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseList-models_Submission"
                        }
                    }
                }
            }
        },
        "/submission/{id}": {
            "get": {
                "description": "get a submission",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Submission"
                ],
                "summary": "Get a submission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Submission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Submission"
                        }
                    }
                }
            }
        },
        "/task/{taskId}": {
            "get": {
                "description": "The task represents a background job that can be run by the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Get a task by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The task ID",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSingle-services_TaskResponse"
                        }
                    }
                }
            }
        },
        "/task/{taskId}/stream": {
            "get": {
                "description": "The task represents a background job that can be run by the system. This endpoint streams the response",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Get a task by ID but stream the response",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The task ID",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/callback": {
            "get": {
                "description": "Handle the OAuth2 callback",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Handle the OAuth2 callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The code from the OAuth2 provider",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The state",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The base URL",
                        "name": "baseURL",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSingle-routes_RedirectResponse"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "get": {
                "description": "Redirect to the OAuth2 login",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Redirect to the OAuth2 login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The callback URL",
                        "name": "callback",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseSingle-routes_RedirectResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.MediaType": {
            "type": "string",
            "enum": [
                "ARTICLE",
                "BITMAP",
                "AUDIO",
                "VIDEO",
                "PDF"
            ],
            "x-enum-varnames": [
                "MediaTypeArticle",
                "MediaTypeImage",
                "MediaTypeAudio",
                "MediaTypeVideo",
                "MediaTypePDF"
            ]
        },
...
        "models.ResponseList-models_Submission": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Submission"
                    }
                },
                "next": {
                    "type": "string"
                },
                "prev": {
                    "type": "string"
                }
            }
        },
...
        "models.ResponseSingle-routes_RedirectResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/routes.RedirectResponse"
                }
            }
        },
...
        "models.ResponseSingle-services_TaskResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/services.TaskResponse"
                }
            }
        },
...
        "models.ScoreType": {
            "type": "number",
            "enum": [
                100
            ],
            "x-enum-varnames": [
                "MAXIMUM_EVALUATION_SCORE"
            ]
        },
        "models.Submission": {
            "type": "object",
            "properties": {
                "assignmentCount": {
                    "description": "The number of times the submission has been assigned to the juries",
                    "type": "integer"
                },
                "author": {
                    "description": "The Actual Author in the Wikimedia",
                    "type": "string"
                },
                "bitrate": {
                    "description": "in kbps",
                    "type": "integer"
                },
                "campaignId": {
                    "type": "string"
                },
                "createdAtServer": {
                    "type": "string"
                },
                "creditHTML": {
                    "type": "string"
                },
                "currentRoundId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "distributionTaskId": {
                    "type": "string"
                },
                "duration": {
                    "description": "in milliseconds",
                    "type": "integer"
                },
                "evaluationCount": {
                    "description": "The number of times the submission has been evaluated by the juries",
                    "type": "integer"
                },
                "height": {
                    "type": "integer"
                },
                "importTaskId": {
                    "type": "string"
                },
                "license": {
                    "type": "string"
                },
                "mediatype": {
                    "$ref": "#/definitions/models.MediaType"
                },
                "metadata": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "pageId": {
                    "type": "integer"
                },
                "participantId": {
                    "type": "string"
                },
                "resolution": {
                    "type": "integer"
                },
                "score": {
                    "description": "The Average Score of the submission",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ScoreType"
                        }
                    ]
                },
                "size": {
                    "description": "in bytes",
                    "type": "integer"
                },
                "submissionId": {
                    "type": "string"
                },
                "submittedAt": {
                    "type": "string"
                },
                "submittedById": {
                    "description": "The User who submitted the article on behalf of the participant",
                    "type": "string"
                },
                "thumbheight": {
                    "type": "integer"
                },
                "thumburl": {
                    "type": "string"
                },
                "thumbwidth": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
...
        "models.TaskStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "success",
                "failed"
            ],
            "x-enum-varnames": [
                "TaskStatusPending",
                "TaskStatusRunning",
                "TaskStatusSuccess",
                "TaskStatusFailed"
            ]
        },
        "models.TaskType": {
            "type": "string",
            "enum": [
                "import.commons",
                "import.previous.round",
                "distribute.evaluations"
            ],
            "x-enum-varnames": [
                "TaskTypeImportFromCommons",
                "TaskTypeImportFromPreviousRound",
                "TaskTypeDistributeEvaluations"
            ]
        },
        "routes.RedirectResponse": {
            "type": "object",
            "properties": {
                "redirect": {
                    "description": "Redirect is the URL to redirect to",
                    "type": "string"
                }
            }
        },
...
        "services.TaskResponse": {
            "type": "object",
            "properties": {
                "campaignId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdById": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "failedCount": {
                    "type": "integer"
                },
                "failedIds": {
                    "type": "object"
                },
                "remainingCount": {
                    "type": "integer"
                },
                "roundId": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.TaskStatus"
                },
                "successCount": {
                    "type": "integer"
                },
                "taskId": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.TaskType"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        }
    }
}